Apply suggestions from Felix's code review

+++ src/MModuleEnergyCalibration.cxx
@@ -166,7 +166,7 @@ bool MModuleEnergyCalibration::AnalyzeEvent(MReadOutAssembly* Event)
     // calibrated energy is not trustworthy. The hit is kept, it is only marked.
     if (SH->GetADCUnits() > m_HighADCThreshold) {
       if (g_Verbosity >= c_Warning) {
-        cout << m_XmlTag << ": Warning: High ADC value " << SH->GetADCUnits() << " for read-out element " << R << endl;
+        cout << m_XmlTag << ": Warning: High ADC value " << (int) SH->GetADCUnits() << " for read-out element " << R << endl;
       }
       Event->SetHighADC_QualityFlag("High ADC value " + to_string((int) SH->GetADCUnits()) + " for " + R.ToString().Data());
     }
